requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
aspell-python-py3==1.15
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
bs4==0.0.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
dnspython==2.0.0
email-validator==2.1.1
flask==3.0.3
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.7
flask-sqlalchemy==3.1.1
flask-wtf==1.2.1
frozenlist==1.4.1
geoip2==4.8.0
greenlet==3.0.3
idna==3.7
importlib-metadata==7.1.0
importlib-resources==6.4.0
ipwhois==1.2.0
itsdangerous==2.2.0
Jinja2==3.1.3
jwt==1.3.1
langid==1.1.6
Mako==1.3.3
MarkupSafe==2.1.5
maxminddb==2.6.1
multidict==6.0.5
mysqlclient==2.2.4
numpy==1.24.4
pyasn==1.6.2
pycparser==2.22
PyJWT==2.8.0
pyphen==0.15.0
requests==2.31.0
soupsieve==2.5
SQLAlchemy==2.0.29
SQLAlchemy-Utils==0.41.2
textstat==0.7.3
typing-extensions==4.11.0
urllib3==2.2.1
werkzeug==3.0.2
whois==0.9.27
wtforms==3.1.2
yarl==1.9.4
zipp==3.18.1
